// src/common/enums/dto/stock.enum.ts
export enum STOCK_DTO_MESSAGE {
  // change type
  CHANGE_TYPE_IS_NOT_EMPTY = 'กรุณาเลือกประเภทการเปลี่ยนแปลง',
  INVALID_CHANGE_TYPE = 'ประเภทการเปลี่ยนแปลงไม่ถูกต้อง',

  // note
  INVALID_LENGTH_NOTE = 'โน้ตต้องมีความยาวระหว่าง 1 ถึง 255 ตัวอักษร',
  NOTE_MUST_BE_STRING = 'โน้ตต้องเป็นข้อความ',
}