// src/common/enums/dto/common.enum.ts
export enum COMMON_DTO {
  UUID_VERSION = '4',

  // quantity
  QTY_IS_NOT_EMPTY = 'กรุณากรอกจำนวน',
  QTY_IS_INTEGER = 'จำนวนต้องเป็นตัวเลขเท่านั้น',
  QTY_IS_POSITIVE = 'จำนวนต้องเป็นตัวเลขที่มากกว่า 0',
}